requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.3.3
autopep8==1.5
boto3==1.14.30
botocore==1.17.30
bottle==0.12.18
bottle-websocket==0.2.9
bson==0.5.8
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorzero==1.1
cycler==0.10.0
dataclasses==0.6
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-cookies-samesite==0.6.6
django-cors-headers==3.2.1
django-decouple==2.1
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djongo==1.3.1
djoser==2.0.3
docutils==0.15.2
dwollav2==2.0.0
Eel==0.12.3
filelock==3.0.12
Flask==1.1.1
future==0.18.2
geographiclib==1.50
geopy==1.21.0
gevent==20.5.0
gevent-websocket==0.10.1
gpiozero==1.5.1
greenlet==0.4.15
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
pandas==1.0.3
Pillow==7.0.0
postgres==3.0.0
psutil==5.6.7
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pycodestyle==2.5.0
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
python-engineio==3.11.2
python-environ==0.4.54
python-socketio==4.4.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scipy==1.4.1
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
timeout==0.1.2
timeout-decorator==0.4.1
twilio==6.44.1
ua-parser==0.10.0
urllib3==1.25.10
virtualenv==20.0.13
Werkzeug==1.0.0
whichcraft==0.6.1
whitenoise==5.0.1
wrapt==1.11.2
